fix(bug): casting exception
class java.lang.Integer cannot be cast to class java.lang.Long
(java.lang.Integer and java.lang.Long are in module java.base of loader
'bootstrap')

+++ src/main/java/com/jhpark/marketing/blog/domain/oauth2/KakaoOAuth2UserInfo.java
@@ -21,7 +21,7 @@ public class KakaoOAuth2UserInfo extends OAuth2UserInfo{
 
   @Override
   public String getId() {
-    return ((Long) super.attributes.get("id")).toString();
+    return ((Integer) super.attributes.get("id")).toString();
   }
 
   @Override
